requirements.txt
langchain==0.0.312
fastapi==0.103.2
python-dotenv==1.0.0
bs4==0.0.1
openai==0.28.1
uvicorn==0.23.2
tiktoken==0.4.0